third_party/WebKit/LayoutTests/external/wpt/dom/events/AddEventListenerOptions-passive-expected.txt

This is a testharness.js-based test.
PASS Supports passive option on addEventListener only
PASS preventDefault should be ignored if-and-only-if the passive option is true
FAIL returnValue should be ignored if-and-only-if the passive option is true assert_equals: Incorrect defaultPrevented for options: {"passive":true} expected false but got true
PASS passive behavior of one listener should be unaffected by the presence of other listeners
PASS Equivalence of option values
Harness: the test ran to completion.
